+++ toolkit/modules/shop_result_explorer/streamlit/shop_result_explorer/pages/shop_scenarios.py
@@ -58,7 +58,7 @@ def run_shop(
         workflowStep=-1,
         functionCallId="_",
     )
-    power_ops_client.v1.upsert(shop_trigger_input)
+    power_ops_client.powermodel.upsert(shop_trigger_input)
     return power_ops_client.cdf.functions.call(
         external_id="shop_trigger",
         data={"shop_trigger_input_instance_external_id": shop_trigger_input.external_id},